Update build scripts

+++ build-fltk.sh
@@ -10,8 +10,9 @@ buildflags="-O2 -fstack-protector -ffunction-sections -fdata-sections -D_FORTIFY
 set -v
 set -e
 
-#test -d fltk || svn co --username '' --password '' "http://seriss.com/public/fltk/fltk/branches/branch-1.3/" fltk
-test -d fltk || git clone -b fltk-1.3.4-1-source https://github.com/darealshinji/fltk-dialog fltk
+if [ ! -d fltk ]; then
+  (git submodule init && git submodule update) || exit
+fi
 
 rm -rf fltk/build
 mkdir -p fltk/build
